requirements.txt
py-cord==2.0.0b7
firebase-admin==5.2.0
python-dotenv==0.20.0
click==8.1.2
humanize==4.0.0